(Agrega) Comandos nuevos
/comunidades - Muestra enlace de github de OpenVE de las comunidades en Telegram

/github - Muestra enlace de github del bot
